==> 2. Feature Engineering-Copy1.ipynb <==
import pandas as pd

!ls

users = pd.read_csv('final_user_data.csv')
browsing_data = pd.read_csv('final_user_browsing_history.csv')
experiences = pd.read_csv('experiences_data_final.csv')
adventures = pd.read_csv('adventures_data_final.csv')
tours = pd.read_csv('tours_data_final.csv')

users.head()

browsing_data

experiences.head()

adventures

tours

tours['tour_data'][0]

# Import the ast module to safely evaluate the string as a Python literal
import ast  

tours['description'] = tours['tour_data'].apply(lambda x: ast.literal_eval(x)['overview']['route_description'])

tours.head()

tours

tours['name'] = tours['name'].str.split(n=1).str[1]
tours

adventures.loc[:, 'price'] = adventures['price'] * 100

adventures

combined = pd.concat([tours, adventures], ignore_index=True)

combined

combined.to_csv('final_tours_and_adventures.csv', index=False)

# Lets Transform the experiences now

experiences.drop(['ratingsAverage', 'signatureSelection'], inplace=True, axis=1)
experiences.head()

categories = pd.read_json('tajriba.categories.json')
personalities = pd.read_json('tajriba.personalities.json')

categories

personalities
