==> 3. Data Preprocessing.ipynb <==
import pandas as pd

!ls

users = pd.read_csv('final_user_data.csv')
browsing_history = pd.read_csv('final_user_browsing_history.csv')
experiences = pd.read_csv('final_experiences_data.csv')
personalities = pd.read_csv('final_personalities_data.csv')
categories = pd.read_csv('final_categories_data.csv')
tours_to_recommend = pd.read_csv('final_tours_and_adventures.csv')

browsing_history.head()

browsing_history.rename(columns={'user_id': 'user', 'experience_id': '_id'}, inplace=True, errors='raise')
browsing_history.head()

merged_data = pd.merge(browsing_history, users, on='user', how='left')
merged_data.head()

merged_data.describe().T

merged_data = pd.merge(merged_data, experiences, on='_id', how='left')
merged_data.columns

merged_data.drop(['_id', 'viewed', 'name_x', 'description_x', 'coverImage_x',
       'featured_x', 'adventureLevel_x', 'price_x', 'categories_x',
       'personality_x', 'name_y', 'description_y', 'coverImage_y',
       'featured_y', 'adventureLevel_y', 'price_y', 'categories_y',
       'personality_y',], inplace=True, axis=1)

merged_data.sample(5)

merged_data.dropna()

merged_data.info()

merged_data.to_csv('final_merged_data.csv', index=False)
