==> 4 More Cleaning.ipynb <==
import pandas as pd

data = pd.read_csv('final_merged_data.csv')

data.head()

data['createdAt'] = pd.to_datetime(data['createdAt'])
data.info()

data = data.dropna()

data.isnull().sum()

# Categorical columns for one-hot encoding
categorical_columns = ['gender', 'accomodation_preference', 'transport_preference', 'featured']

# One-hot encoding
data = pd.get_dummies(data, columns=categorical_columns, drop_first=True)

data.head()

data.info()

# Identify boolean columns
bool_columns = data.select_dtypes(include='bool').columns

# Convert boolean columns to integer
data[bool_columns] = data[bool_columns].astype(int)

# Display the updated DataFrame
data.head()

data.describe().T

data.columns

data.to_csv('final_merged_data.csv', index=False)
